--- human_aware_rl/data/human/anonymized/HAI_questiondata.csv
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:96c81286900d2d47f6a05ea47a13850070c6301e9468456a6ca568a828d1059d
-size 26122
